combined spurious_correlation_handler.py as a class in imagelab.py
